requirements_cloud_funciton.txt
functions-framework
smtplib
flask
google-cloud-secret-manager
email